Split binary patcher into separate class
