+++ OWML.Patcher/OWML.Patcher.csproj
@@ -53,6 +53,7 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BinaryPatcher.cs" />
     <Compile Include="OWPatcher.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="VRPatcher.cs" />
